nonupgradable test to look for tables that use system defined types
System defined types are not guaranteed to have consistent oids when
moving from on version of postgres to the next. Unfortunately, there are
quite a few types that got their OID changed from GPDB6 (postgres 9.4)
to GPDB7 (12.12). The query looking user table that use these types had
to be rewritten for upgrade because the recursive query looking for
these types of relations contained a self reference in a subquery. This
specific type of query is disabled in 6x so it was rewritten in plpgsql.

gpdb reference commit:
https://github.com/greenplum-db/gpdb/pull/16820